Controller and device (robot) are now fully separated
Some polishing of YumiCoordinaredRobotState must be done, but overall the current APIs are ok. This commit adds the `AbstractDevice` class, which is in charge of communicating with the hardware (or simulated) device. The controller simply asks the device to read or send, while it can focus on the policy. This also means that controllers are now objects capable of controlling generic states, and are not bound to a specific device anymore.